Move is_create_vm to fix circular imports
